+++ inst/doc/introduction.R
@@ -4,11 +4,12 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
-if (!requireNamespace("rdflib", quietly = TRUE)) {
+if (!requireNamespace("rdflib", 
+                      quietly = TRUE)) {
   stop("Please install 'rdflib' to run this vignette.")
 }
 
-## -----------------------------------------------------------------------------
+## ----definegdpdataset---------------------------------------------------------
 library(dataset)
 
 gdp <- defined(
@@ -43,7 +44,7 @@ small_dataset <- dataset_df(
 
 small_dataset
 
-## -----------------------------------------------------------------------------
+## ----dublincoremetadata-------------------------------------------------------
 as_dublincore(small_dataset)
 
 ## ----triplesexample-----------------------------------------------------------
@@ -53,8 +54,10 @@ triples <- dataset_to_triples(small_dataset,
 triples
 
 ## ----ntexample----------------------------------------------------------------
-mycon <- tempfile("my_dataset", fileext = "nt")
-my_description <- describe(x = small_dataset, con = mycon)
+mycon <- tempfile("my_dataset", 
+                  fileext = "nt")
+my_description <- describe(x = small_dataset, 
+                           con = mycon)
 
 # Only three statements are shown:
 readLines(mycon)[c(4, 8, 12)]
@@ -62,4 +65,17 @@ readLines(mycon)[c(4, 8, 12)]
 ## ----provenancexample---------------------------------------------------------
 ## Show two lines of provenance:
 provenance(small_dataset)[c(6, 7)]
+
+## ----smalldf------------------------------------------------------------------
+small_df <- as.data.frame(small_dataset, 
+              strip_attributes = FALSE)
+
+attr(small_dataset, "subject")
+
+## ----smalltbl-----------------------------------------------------------------
+small_tbl <- as_tibble(
+  small_dataset, 
+  strip_attributes = TRUE)
+
+small_tbl